I put the recipe list in its own page and the recipe detail screen in its own page. the home screen now navigates to the recipe search once you type in a food.
