+++ recipeapp/lib/pages/home.dart
@@ -1,10 +1,7 @@
-import 'dart:convert';
+// lib/pages/home.dart
 import 'package:flutter/material.dart';
-import 'package:http/http.dart' as http;
 import 'package:google_fonts/google_fonts.dart';
-
-const appId = '630a63de';
-const appKey = '2ebdde9075b8b570315e2734bd35f0ce';
+import 'RecipeSearchList.dart'; 
 
 class MyHomePage extends StatefulWidget {
   const MyHomePage({super.key, required this.title});
@@ -16,26 +13,7 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  List<Map<String, dynamic>> recipes = [];
-
-  Future<void> fetchRecipes(String query) async {
-    final url = Uri.parse(
-        'https://api.edamam.com/search?q=$query&app_id=$appId&app_key=$appKey');
-
-    final response = await http.get(url);
-
-    if (response.statusCode == 200) {
-      final data = json.decode(response.body);
-
-      setState(() {
-        recipes = (data['hits'] as List)
-            .map((hit) => hit['recipe'] as Map<String, dynamic>)
-            .toList();
-      });
-    } else {
-      print('Failed to fetch recipes: ${response.statusCode}');
-    }
-  }
+  TextEditingController _controller = TextEditingController();
 
   @override
   Widget build(BuildContext context) {
@@ -52,8 +30,10 @@ class _MyHomePageState extends State<MyHomePage> {
                 : screenHeight * 0.3;
 
             return Column(
-              mainAxisAlignment: MainAxisAlignment.center, // Center all elements vertically
-              crossAxisAlignment: CrossAxisAlignment.center, // Center all elements horizontally
+              mainAxisAlignment:
+                  MainAxisAlignment.center, // Center all elements vertically
+              crossAxisAlignment:
+                  CrossAxisAlignment.center, // Center all elements horizontally
               children: [
                 // Cook Logo Section
                 Image.asset(
@@ -84,7 +64,17 @@ class _MyHomePageState extends State<MyHomePage> {
                     child: Padding(
                       padding: const EdgeInsets.symmetric(horizontal: 16.0),
                       child: TextField(
-                        onSubmitted: fetchRecipes,
+                        controller: _controller,
+                        onSubmitted: (query) {
+                          if (query.isNotEmpty) {
+                            Navigator.push(
+                              context,
+                              MaterialPageRoute(
+                                builder: (context) => RecipeSearchList(query: query),
+                              ),
+                            );
+                          }
+                        },
                         decoration: InputDecoration(
                           labelText: 'Enter a recipe name',
                           labelStyle: TextStyle(
@@ -107,32 +97,6 @@ class _MyHomePageState extends State<MyHomePage> {
                     ),
                   ),
                 ),
-
-                // List of recipes
-                Expanded(
-                  child: ListView.builder(
-                    itemCount: recipes.length,
-                    itemBuilder: (context, index) {
-                      var recipe = recipes[index];
-                      var label = recipe['label'];
-                      var imageUrl = recipe['image'];
-                      return Card(
-                        margin: const EdgeInsets.all(8.0),
-                        child: ListTile(
-                          leading: Image.network(imageUrl, width: 50, height: 50, fit: BoxFit.cover),
-                          title: Text(
-                            label,
-                            style: GoogleFonts.dancingScript(
-                              fontSize: screenWidth * 0.05,
-                              fontWeight: FontWeight.bold,
-                              color: Color(0xFF7EA16B),
-                            ),
-                          ),
-                        ),
-                      );
-                    },
-                  ),
-                ),
               ],
             );
           },
@@ -141,3 +105,4 @@ class _MyHomePageState extends State<MyHomePage> {
     );
   }
 }
+
